requirements.txt
asgiref==3.2.10
certifi==2018.11.29
Django==2.2.10
django-ajax-selects==1.3.5
django-crispy-forms==1.7.2
django-debug-toolbar==2.1
django-dynamic-formsets==0.0.8
django-extensions==2.2.7
django-widget-tweaks==1.4.8
Pillow==7.1.2
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
Werkzeug==1.0.0
wincertstore==0.2
